Give the fault guard a VC6 spelling, and stop using long long
oracle_fault_guard.cpp was written for GCC in two places that have no MSVC
equivalent by accident of syntax rather than capability.

Its thunk is a top-level `asm` block in AT&T, which VC6 answers with
`C2290: C++ 'asm' syntax ignored`. The same three instructions - cld, call,
hlt - become a `__declspec(naked)` function with an Intel body, where the
leading underscore the AT&T version spells out is the compiler's job. This is
scaffolding and not a recovered body, so the bar on inline assembly in
recovered code does not apply: nothing here is copied from the original.

`AddVectoredExceptionHandler` shipped with Windows XP and VC6's Platform SDK
predates it. The entry point is in every kernel32 this runs against including
wine's, so only the declaration was missing. It is declared at its one use
site rather than in vc6_compat.h, because that header is included before
<windows.h> and LONG and EXCEPTION_POINTERS do not exist yet at that point.

`long long` reached wave.cpp again through a body integrated since the
generator was fixed; `int64_t` is the same type with a name VC6 has.

+++ src/wave.cpp
@@ -661,7 +661,7 @@ void Wave::set_volume(int a1) {
         // The original loads the group dword zero-extended through a 64-bit
         // fild, so the scale is the UNSIGNED value of the table entry.
         const double group = static_cast<double>(WaveDeviceGroupVolumes[group_slot_ * 6]);
-        level = static_cast<int>(static_cast<long long>(
+        level = static_cast<int>(static_cast<int64_t>(
             static_cast<double>(static_cast<int>(vol)) * (1.0 / 127.0) *
             group));
     }
